fix update static shell script

+++ .github/update-nwi-static.sh
@@ -1,24 +1,27 @@
 #!/bin/bash
 
 set -e
-# update.nwi.sh {static-type} {source path}
+# update-nwi-static.sh {static-type} {source path}
 
 type=$1
 source_path=$2
 
 echo "received $type from $source_path"
 
-services=$(docker service ls --format "{{.Name}}" | grep nwi-)
+services=$(ls /var/instances/*/ -d | grep nwi-)
 
 for service in ${services[@]}
 do
-    dist_path="/var/instances/$service/public/$type"
+    dist_path="${service}public/$type"
 
-    # rm static
-    rm -fr $dist_path/*
-    echo "remove static $type for $service"
+    if [ ${source_path::-1} != $dist_path ]
+    then
+      # rm static
+      rm -fr $dist_path/*
+      echo "remove static $type for $service"
 
-    # copy static
-    cp -r "$source_path*" $dist_path 
-    echo "copy static $type for $service"
+      # copy static
+      cp -r ${source_path}* $dist_path
+      echo "copy static $type for $service"
+    fi
 done
